Move ip checker to utils
